+++ HISTORY.md
@@ -1,5 +1,11 @@
 # History
 
+## 1.1.0 (2021-11-09)
+
+* Multi-repository support (Azure Docs, Azure SQL, Azure Quantum)
+* Review home page
+* Navigation breadcrumb
+
 ## 1.0.0 (2021-11-03)
 
 * Use a GitHub oAuth token to increase the rate limit and the number of results.